+++ scripts/macros/2024/spells/dissolution.js
@@ -1,0 +1,5 @@
+export let dissolution = {
+    name: 'Dissolution',
+    version: '1.3.159',
+    rules: 'modern'
+};